mv almost 500 lines from lightclient impl into new integration_helpers mod
